# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/batches": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch"
                ],
                "summary": "批次列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Batch"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch"
                ],
                "summary": "创建批次",
                "parameters": [
                    {
                        "description": "批次信息",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.batchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Batch"
                        }
                    }
                }
            }
        },
        "/batches/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch"
                ],
                "summary": "批次详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "批次 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Batch"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Batch"
                ],
                "summary": "更新批次",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "批次 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "批次信息",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.batchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Batch"
                        }
                    }
                }
            }
        },
...
        "/tasks": {
            "get": {
                "description": "查询任务列表，可分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ListTasksResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{task_id}": {
            "get": {
                "description": "根据 task_id 查询单个任务",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "任务详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TaskDetailResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{task_id}/logs": {
            "get": {
                "description": "查询某个任务的日志",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "任务日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "每页条数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TaskLogResponse"
                        }
                    }
                }
            }
        },
        "/templates": {
            "get": {
                "description": "查询所有任务模板",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "模板列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ParamTemplate"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "创建模板",
                "parameters": [
                    {
                        "description": "模板信息",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.templateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ParamTemplate"
                        }
                    }
                }
            }
        },
        "/templates/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "更新模板",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "模板 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "模板信息",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.templateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ParamTemplate"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Template"
                ],
                "summary": "删除模板",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "模板 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.ListTasksResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TaskItem"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/api.TaskPagination"
                }
            }
        },
...
        "api.TaskDetailResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "duration_secs": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "log_summary": {
                    "type": "string"
                },
                "queued_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "trigger": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.TaskItem": {
            "type": "object",
            "properties": {
                "finished_at": {
                    "type": "string"
                },
                "queued_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "trigger": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.TaskLogItem": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "api.TaskLogResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TaskLogItem"
                    }
                }
            }
        },
        "api.TaskPagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.batchRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "refreshed": {
                    "type": "boolean"
                }
            }
        },
...
        "api.templateRequest": {
            "type": "object",
            "required": [
                "name",
                "task_type"
            ],
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                }
            }
        },
...
        "models.Batch": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastRefreshedAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ParamTemplate": {
            "type": "object",
            "properties": {
                "configJSON": {
                    "type": "object"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "taskType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}